Peering network preview
This is the great renewal of those that I wrote.
I'm added NMDC protocol support.
This means that the data can come from a peer network or cloud. This is
the idea of creating this program.
This Preview version. And then there's a lot of things to be done. In
the near versions will add support for the distribution of traffic to
other users. Addition of velocities http, ftp and nmdc! This will
significantly increase the speed of downloading. And this is not the
limit! Mathematical calculations will help to increase the data transfer
rate of 10 or 100 times !!